+++ src/utils/common.js
@@ -1,4 +1,9 @@
 import dayjs from 'dayjs';
+import {
+  MAX_DAYS_GAP,
+  MAX_MINUTES,
+  MIN_VALUE
+} from './consts.js';
 
 export const getRandomInteger = (a = 0, b = 1) => {
   const lower = Math.ceil(Math.min(a, b));
@@ -19,4 +24,15 @@ export const generateDate = () => {
   const daysGap = getRandomInteger(MIN_VALUE, MAX_DAYS_GAP);
 
   return dayjs().add(-daysGap, 'day').toDate();
+};
+
+export const generateDuration = () => {
+  const hours = getRandomInteger();
+  const minutes = getRandomInteger(MIN_VALUE, MAX_MINUTES);
+
+  let durations = hours ? `${hours}h ` : '';
+
+  durations += `${minutes}m`;
+
+  return durations;
 };
