+++ src/utils/task.js
@@ -1,16 +1,11 @@
 import dayjs from 'dayjs';
 import { getRandomPositiveFloat } from './common.js';
+import {
+  Date,
+  MIN_VALUE,
+  MAX_RATING
+} from './consts.js';
 
-export const generateDuration = () => {
-  const hours = getRandomInteger();
-  const minutes = getRandomInteger(MIN_VALUE, MAX_MINUTES);
-
-  let durations = hours ? `${hours}h ` : '';
-
-  durations += `${minutes}m`;
-
-  return durations;
-};
 
 export const getRating = () => getRandomPositiveFloat(MIN_VALUE, MAX_RATING);
