+++ src/view/popup-film-view.js
@@ -1,4 +1,4 @@
-import { datePopup } from '../utils/helpers.js';
+import { datePopup } from '../utils/task.js';
 import AbstractView from './abstract-view.js';
 import { createlsItemButton } from './film-card-view.js';
 import { Selectors } from '../utils/consts.js';
